+++ src/PageUtils/PageNavigation.php
@@ -57,8 +57,9 @@ class PageNavigation
      * @throws Exception\CommunicationException
      * @throws Exception\CommunicationException\CannotReadResponse
      * @throws Exception\CommunicationException\InvalidResponse
+     * @throws Exception\NoResponseAvailable
      */
-    public function __construct(Page $page, string $url, ?string $preScript = NULL)
+    public function __construct(Page $page, string $url)
     {
 
         // make sure latest loaderId was pulled
@@ -66,13 +67,6 @@ class PageNavigation
 
         // get previous loaderId for the navigation watcher
         $this->previousLoaderId = $page->getFrameManager()->getMainFrame()->getLatestLoaderId();
-
-
-        if ($preScript !== NULL){
-            $page->getSession()->sendMessageSync(
-                new Message('Page.addScriptToEvaluateOnNewDocument', ['source'=>$preScript])
-            );
-        }
 
         // send navigation message
         $this->navigateResponseReader = $page->getSession()->sendMessage(